refactor: modularize bot entry
