+++ bot/src/config.js
@@ -1,0 +1,20 @@
+const path = require('path');
+const dotenv = require('dotenv');
+
+dotenv.config({ path: path.join(__dirname, '..', '.env') });
+
+module.exports = {
+  port: process.env.PLUGIN_PORT || 3000,
+  discord: {
+    token: process.env.DISCORD_BOT_TOKEN,
+    clientId: process.env.DISCORD_CLIENT_ID,
+    apolloBotId: process.env.APOLLO_BOT_ID
+  },
+  db: {
+    host: process.env.DB_HOST,
+    user: process.env.DB_USER,
+    password: process.env.DB_PASSWORD,
+    name: process.env.DB_NAME
+  }
+};
+
