+++ bot/src/db/index.js
@@ -2,12 +2,12 @@ const mysql = require('mysql2/promise');
 
 let pool;
 
-async function init() {
+async function init(config) {
   pool = mysql.createPool({
-    host: process.env.DB_HOST,
-    user: process.env.DB_USER,
-    password: process.env.DB_PASSWORD,
-    database: process.env.DB_NAME,
+    host: config.host,
+    user: config.user,
+    password: config.password,
+    database: config.name,
     waitForConnections: true,
     connectionLimit: 10,
     queueLimit: 0
